dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev-requirements.txt dev-requirements.in
#

boto3==1.3.1
botocore==1.4.16          # via boto3
docutils==0.12            # via botocore
greenlet==0.4.9           # via neovim
jmespath==0.9.0           # via boto3, botocore
msgpack-python==0.4.7
neovim==0.1.7
python-dateutil==2.5.3    # via botocore
pyyaml==3.11
six==1.10.0               # via python-dateutil
